move logger initialization to main process
