+++ packages/desktop/lib/electron/interfaces/version-details.interface.ts
@@ -1,0 +1,7 @@
+export interface IVersionDetails {
+    upToDate: boolean
+    currentVersion: string
+    newVersion: string
+    newVersionReleaseDate: Date
+    changelog: string
+}
