+++ packages/desktop/lib/electron/preloads/preload.ts
@@ -6,18 +6,11 @@
  * https://www.electronjs.org/docs/latest/tutorial/sandbox
  */
 
-import fs from 'fs'
 import { ipcRenderer, contextBridge } from 'electron'
-
-import * as IotaSdk from '@iota/sdk'
-import type { ILoggerConfig } from '@iota/sdk/out/types'
 
 import ElectronApi from '../apis/electron.api'
 import LedgerApi from '../apis/ledger.api'
 import WalletApi from '../apis/wallet.api'
-
-const DAY_IN_MILLISECONDS = 1000 * 60 * 60 * 24
-const DAYS_TO_KEEP_LOGS = 30
 
 // Hook the error handlers as early as possible
 window.addEventListener('error', handleErrorEvent)
@@ -30,17 +23,6 @@ window.addEventListener('DOMContentLoaded', () => {
         ipcRenderer.send('dom-content-loaded')
     }, 200)
 })
-
-try {
-    if (process.env.STAGE !== 'prod') {
-        void ipcRenderer.invoke('get-path', 'userData').then(async (baseDir) => {
-            const logDir = prepareLogDirectory(baseDir)
-            await getVersionAndInitLogger(logDir)
-        })
-    }
-} catch (err) {
-    console.error('[Preload Context] Error:', err)
-}
 
 try {
     // contextBridge doesn't allow passing custom properties & methods on prototype chain
@@ -72,42 +54,4 @@ function handleUnhandledRejectionEvent(event: PromiseRejectionEvent): void {
     void ipcRenderer.invoke('handle-error', '[Preload Context] Unhandled Rejection', event.reason)
     event.preventDefault()
     console.error(event.reason)
-}
-
-function prepareLogDirectory(baseDir: string): string {
-    const logDir = `${baseDir}/logs`
-    if (!fs.existsSync(logDir)) {
-        fs.mkdirSync(logDir)
-    }
-    return logDir
-}
-
-async function getVersionAndInitLogger(logDir: string): Promise<void> {
-    const versionDetails = await ipcRenderer.invoke('get-version-details')
-    const today = new Date().toISOString().slice(0, 16).replace('T', '-').replace(':', '-')
-    const loggerOptions: ILoggerConfig = {
-        colorEnabled: true,
-        name: `${logDir}/wallet-v${versionDetails.currentVersion}-d${today}.log`,
-        levelFilter: 'debug',
-        targetExclusions: ['h2', 'hyper', 'rustls', 'message_handler'],
-    }
-    IotaSdk.initLogger(loggerOptions)
-
-    deleteOldLogs(logDir, versionDetails.currentVersion)
-}
-
-function deleteOldLogs(path: string, currentVersion: string): void {
-    const files = fs.readdirSync(path)
-    files.forEach((file) => {
-        const filePath = path + '/' + file
-        const stat = fs.statSync(filePath)
-
-        const isOlderThan30Days =
-            new Date().getTime() - new Date(stat.mtime).getTime() > DAYS_TO_KEEP_LOGS * DAY_IN_MILLISECONDS
-        const version = file.match(/wallet-v((\w*.)*)-d((\w*.)*).log/)?.[1]
-        const isDifferentVersion = version !== currentVersion
-        if (isDifferentVersion || isOlderThan30Days) {
-            fs.unlinkSync(filePath)
-        }
-    })
 }
